backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.3
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn==22.0.0